Added ViewGrid to its own file.
